+++ doc/whatsnew.rst
@@ -10,6 +10,8 @@ A catalog of new features, improvements, and bug-fixes in each release.
 v0.9.dev
 --------
 
+- Add conjugate transpose operator :func:`pyriemann.utils.base.ctranspose` for real- and complex-valued ndarrays. :pr:`348` by :user:`qbarthelemy`
+
 - Add :class:`pyriemann.embedding.TSNE`, a Riemannian t-SNE implementation
   and update example comparing embeddings. :pr:`347` by :user:`thibaultdesurrel`